extract track layer to separate component
